FW-5472 Extract immersion toggle into a reuseable component
