+++ src/components/ImmersionToggle/index.js
@@ -1,0 +1,3 @@
+import ImmersionToggle from 'components/ImmersionToggle/ImmersionTogglePresentation'
+
+export default ImmersionToggle
